% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_quant.R
\name{decompress_quant}
\alias{decompress_quant}
\title{decompress the fetched quantification result of a
specific dataset and record the path
as the quant_path field of the returned processed_quant list
This function must be run after \code{fetch_quant()}}
\usage{
decompress_quant(
  processed_quant,
  quant_dir = "processed_quant",
  quant_path_name = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{processed_quant}{a list recording the details of
a dataset. Initialized by \code{init_processed_quant(dataset_id)}}

\item{quant_dir}{a string to a path where the fetched tar files will
be stored. It will be created if does not exist.}

\item{quant_path_name}{a string indicates the name of the directory
that will be used for storing the quantification result. If NULL,
the dataset_id will be used as the file name.}

\item{force}{logic whether to proceed if the tar file exists.}

\item{quiet}{logical whether to display no messages}
}
\description{
decompress the fetched quantification result of a
specific dataset and record the path
as the quant_path field of the returned processed_quant list
This function must be run after \code{fetch_quant()}
}
