% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessedQuant.R
\name{FDL}
\alias{FDL}
\title{fetch, decompress and load the quantification result of a
specific dataset as a SingleCellExperiment object
and store it in the sce slot of the returned
processed_quant list.}
\usage{
FDL(
  dataset_id,
  tar_dir = "quant_tar",
  tar_name = NULL,
  quant_dir = "processed_quant",
  quant_path_name = NULL,
  output_format = "scRNA",
  nonzero = FALSE,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{dataset_id}{the id of an available dataset.
Run \code{print_available_datasets()} for all available
datasets.}

\item{tar_dir}{a string to a path where the fetched tar files will
be stored. It will be created if does not exist.}

\item{tar_name}{a string indicates the name of the tar file. If NULL,
the dataset_id will be used as the file name.}

\item{quant_dir}{a string to a path where the fetched tar files will
be stored. It will be created if does not exist.}

\item{quant_path_name}{a string indicates the name of the directory
that will be used for storing the quantification result. If NULL,
the dataset_id will be used as the file name.}

\item{output_format}{the format of the returned SingleCellExperiement
object. It will be passed to \code{\link[fishpond]{loadFry}}
as the \code{outputFormat} parameter.}

\item{nonzero}{It will be passed to \code{\link[fishpond]{loadFry}}
as the \code{nonzero} parameter.}

\item{force}{logic whether to proceed if the files exist.}

\item{quiet}{logical whether to display no messages}
}
\value{
A complete ProcessedQuant class object with valid
\code{tar_path}, \code{quant_path} and \code{sce} slots.
}
\description{
fetch, decompress and load the quantification result of a
specific dataset as a SingleCellExperiment object
and store it in the sce slot of the returned
processed_quant list.
}
