% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessedQuant.R
\name{fetch_quant}
\alias{fetch_quant}
\title{fetch the compressed quantification result of a
specific dataset and stroe the path to the tar file
in the tar_path slot of the returned processed_quant list.
This function must be run after \code{ProcessedQuant()}}
\usage{
fetch_quant(
  processed_quant,
  tar_dir = "quant_tar",
  tar_name = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{processed_quant}{a \code{ProcessedQuant} class object
recording the details of a dataset, which is nitialized by
running \code{ProcessedQuant(dataset_id)}}

\item{tar_dir}{a string to a path where the fetched tar files will
be stored. It will be created if does not exist.}

\item{tar_name}{a string indicates the name of the tar file. If NULL,
the dataset_id will be used as the file name.}

\item{force}{logic whether to proceed if the tar file exists.}

\item{quiet}{logical whether to display no messages}
}
\value{
A ProcessedQuant class object with a updated \code{tar_path} slot.
}
\description{
fetch the compressed quantification result of a
specific dataset and stroe the path to the tar file
in the tar_path slot of the returned processed_quant list.
This function must be run after \code{ProcessedQuant()}
}
