% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_quant.R
\name{fetch_tar}
\alias{fetch_tar}
\title{fetch the compressed quantification result of a 
specific dataset and stroe the path to the tar file 
in the tar_path field of the returned processed_quant list.
This function must be run after \code{init_processed_quant()}}
\usage{
fetch_tar(
  processed_quant,
  tar_dir = "quant_tar",
  file_name = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{processed_quant}{a list recording the details of 
a dataset. Initialized by \code{init_processed_quant(dataset_id)}}

\item{tar_dir}{a string to a path where the fetched tar files will
be stored. It will be created if does not exist.}

\item{file_name}{a string indicates the name of the tar file. If NULL,
the dataset_id will be used as the file name.}

\item{force}{logic whether to proceed if the tar file exists.}

\item{quiet}{logical whether to display no messages}
}
\description{
fetch the compressed quantification result of a 
specific dataset and stroe the path to the tar file 
in the tar_path field of the returned processed_quant list.
This function must be run after \code{init_processed_quant()}
}
