% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_quant.R
\name{init_processed_quant}
\alias{init_processed_quant}
\title{initiate the processed_quant list used for recording the 
information of a specific dataset}
\usage{
init_processed_quant(dataset_id)
}
\arguments{
\item{dataset_id}{the id of an available dataset. 
Run \code{print_available_datasets()} for all available
datasets.}
}
\description{
initiate the processed_quant list used for recording the 
information of a specific dataset
}
