% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load10x.R
\name{load10x}
\alias{load10x}
\title{Download and load 10x datasets}
\usage{
load10x(
  dataset_ids,
  output_dir = "processed_data_quant",
  force = FALSE,
  delete_tar = TRUE,
  output_format = "scRNA",
  nonzero = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{dataset_ids}{integer scalar or vector providing the id of the
dataset(s) to be downloaded and processed. See \code{\link{fetch_processed_quant}}
for details.}

\item{output_dir}{path to the folder that will
be used to store downloaded datasets.
See \code{\link{fetch_processed_quant}} for details.}

\item{force}{logical whether to force re-downloading the existing datasets.
See \code{\link{fetch_processed_quant}} for details.}

\item{delete_tar}{logical whether to delete the intermediate 
compressed datasets.
See \code{\link{fetch_processed_quant}} for details.}

\item{output_format}{can be \emph{either} a valid
\code{\link[fishpond]{loadFry}} \code{outputFormat}
or a list of it. If it is a list, each item in the list
corresponds to a queried dataset and
should be a valid loadFry-\code{outputFormat}.
Each queried dataset will be loaded according to
the corresponding \code{outputFormat} in the list.}

\item{nonzero}{Similar with \code{outputFormat}, it can be
either a valid \code{\link[fishpond]{loadFry}} \code{nonzero}
parameter or a list of it. If a list, the queried datasets
will be processed using the corresponding \code{nonzero}.}

\item{quiet}{logical whether to display no messages}
}
\value{
If a single dataset is queried, a SingleCellExperiment
object will be returned. If multiple datasets are queried,
a list of SingleCellExperiment objects will be returned, each
for a queried dataset.
}
\description{
Download and load preprocessed 10x datasets as SingleCellExperiment objects.
}
\details{
This function is a thin wrapper of the \code{\link[fishpond]{loadFry}}
function and \code{\link{fetch_processed_quant}} function. This function will
download the preprocessed 10x datasets from a remote host and load
them into R as a SingleCellExperiment object or a list of it if
querying multiple datasets. To list all available datasets,
simply run \code{fetch_processed_quant()}
This function takes the combination
of the parameters of the \code{\link[fishpond]{loadFry}}
function and \code{\link{fetch_processed_quant}} function. For
each of the \code{\link[fishpond]{loadFry}} parameters,
a list of that parameter can be specified, each for a
queried dataset.
}
\examples{

\dontrun{
library(roe)
# run the function
# The four way to define output_format
# are the same.
load10x(dataset_ids = c(1, 2),
        output_dir = "10x_datasets",
        force = FALSE,
        delete_tar = TRUE,
        output_format = "scRNA",
#         output_format = list("scRNA", "scRNA"),
#         output_format = list(list(counts = c("S", "A")),
#                               list(counts = c("S", "A"))
#                              ),
#         output_format = list(counts = c("S", "A")),
        nonzero = FALSE,
        quiet = FALSE
)
}

}
