% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_processed_quant.R
\name{load_processed_quant}
\alias{load_processed_quant}
\title{Fetch and load 10x datasets}
\usage{
load_processed_quant(
  dataset_ids = c(),
  fetch_dir = "processed_quant",
  force = FALSE,
  delete_tar = FALSE,
  output_format = "scRNA",
  nonzero = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{dataset_ids}{integer scalar or vector providing the id of the
dataset(s) to be Fetched and processed. See 
\code{\link{fetch_processed_quant}}
for details.}

\item{fetch_dir}{path to the folder where the fetched quantification results
should be stored.
See \code{\link{fetch_processed_quant}} for details.}

\item{force}{logical whether to force re-downloading the existing datasets.
See \code{\link{fetch_processed_quant}} for details.}

\item{delete_tar}{logical whether to delete the compressed datasets after
decompressing. If FALSE, the tar files
 will be stored in a folder called
datasets_tar under the \code{fetch_dir}.}

\item{output_format}{can be \emph{either} a valid
\code{\link[fishpond]{loadFry}} \code{outputFormat}
or a list of it. The \code{names} of the list must
match the \code{dataset_ids}.}

\item{nonzero}{Similar with \code{outputFormat}, it can be
either a valid \code{\link[fishpond]{loadFry}} \code{nonzero}
parameter or a list of it. If a list, The \code{names} of the 
list must match the \code{dataset_ids}.}

\item{quiet}{logical whether to display no messages}
}
\value{
If an empty dataset_ids is provided,
a data frame containing the information of available datasets
will be returned; otherwise, a list of ProcessedQuant class object, 
where each list stores the information of one fetched dataset. 
The `quant_path` slot represent the path to the quantification
result of the fetched dataset; The `sce` slot represent the 
SingleCellExperiment of the quantification result of this dataset.
}
\description{
Fetch and load preprocessed 10x datasets as SingleCellExperiment objects.
}
\details{
This function is a thin wrapper of the \code{\link[fishpond]{loadFry}}
function and \code{\link{fetch_processed_quant}} function. This function will
fetch the quantification result of each dataset in the provided
`dataset_ids` into the `fetch_dir` directory, and load the
quantification result of each fetched dataset into R as a
SingleCellExperiment object. This function returns a list of lists,
in which each list stores the information of a fetched dataset, and
the SingleCellExperiment of that dataset.
To list all available datasets,
simply run \code{load_processed_quant()}.
This function takes the combination
of the parameters of the \code{\link[fishpond]{loadFry}}
function and \code{\link{fetch_processed_quant}} function. For
each of the \code{\link[fishpond]{loadFry}} parameters,
a list of that parameter can be specified, each for a
fetched dataset.
}
\examples{

\dontrun{
library(roe)
# run the function
# The four ways to define output_format
# are the same.
load_processed_quant(dataset_ids = c(1, 3),
        fetch_dir = "processed_quant",
        force = FALSE,
        delete_tar = FALSE,
        output_format = "scRNA",
#         output_format = list("1" = scRNA", "3" = "scRNA"),
#         output_format = list("1" = list(counts = c("S", "A")),
#                               "3" = list(counts = c("S", "A"))
#                              ),
#         output_format = list("counts" = c("S", "A")),
        nonzero = FALSE,
        quiet = FALSE
)


fetched_quant_list$"1"@sce)
fetched_quant_list$"3"@sce)
}

}
