% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_quant.R
\name{load_quant}
\alias{load_quant}
\title{load the fetched quantification result of a
specific dataset as a SingleCellExperiment object
and store it in the sce field of the returned
 processed_quant list.
This function must be run after \code{decompress_tar()}}
\usage{
load_quant(
  processed_quant,
  output_format = "scRNA",
  nonzero = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{processed_quant}{a list recording the details of
a dataset.}

\item{output_format}{the format of the returned SingleCellExperiment
object. It will be passed to \code{\link[fishpond]{loadFry}}
as the \code{outputFormat} parameter.}

\item{nonzero}{It will be passed to \code{\link[fishpond]{loadFry}}
as the \code{nonzero} parameter.}

\item{quiet}{logical whether to display no messages}
}
\description{
load the fetched quantification result of a
specific dataset as a SingleCellExperiment object
and store it in the sce field of the returned
 processed_quant list.
This function must be run after \code{decompress_tar()}
}
