% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_splici_transcriptome.R
\name{make_splici_txome}
\alias{make_splici_txome}
\title{Construct splici transcriptome}
\usage{
make_splici_txome(
  gtf_path,
  genome_path,
  read_length,
  flank_trim_length = 5,
  output_dir,
  file_name_prefix = "transcriptome_splici",
  extra_spliced = NULL,
  extra_unspliced = NULL,
  dedup_seqs = FALSE,
  write_actual_flank = FALSE
)
}
\arguments{
\item{gtf_path}{Character scalar providing the path to a gtf file.}

\item{genome_path}{Character scalar providing the path to a genome fasta
file.}

\item{read_length}{Numeric scalar indicating the read length.}

\item{flank_trim_length}{Numeric scalar giving the flank trimming length. The
final flank length is obtained by subtracting the \code{flank_trim_length}
from the \code{read_length}.}

\item{output_dir}{Character scalar indicating the output directory, where
the splici transcriptome files will be written.}

\item{file_name_prefix}{Character scalar giving the file name prefix of the
generated output files. The derived flank length will be automatically
appended to the provided prefix.}

\item{extra_spliced}{Character scalar providing the path to a fasta file
with additional sequences to include among the spliced ones.}

\item{extra_unspliced}{Character scalar providing the path to a fasta file
with additional sequences to include among the unspliced ones.}

\item{dedup_seqs}{Logical scalar indicating whether or not to remove
duplicate sequences.}

\item{write_actual_flank}{Logical scalar indicating whether or not to write
out the actual flank length (which may be shorter than the indicated one
if the latter would mean going outside the gene boundaries).}
}
\value{
Nothing is returned, but the necessary files for the splici
  transcriptome are written to the designated \code{output_dir}.
}
\description{
Construct the splici transcriptome for alevin-fry.
}
\examples{
## Put a small example here

}
\author{
Dongze He
}
