% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_quant.R
\name{print_available_datasets}
\alias{print_available_datasets}
\title{print the index and the name of available datasets.}
\usage{
print_available_datasets()
}
\description{
print the index and the name of available datasets.
}
