% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessedQuant.R
\docType{class}
\name{ProcessedQuant-class}
\alias{ProcessedQuant-class}
\alias{ProcessedQuant}
\title{The ProcessedQuant class
The ProcessedQuant class is designed to represent the details of
preprocessed dataset from the
\href{https://github.com/COMBINE-lab/10x-requant}{10x-requant}
project. To initialize a ProcessedQuant class, run
\code{ProcessedQuant(dataset_id)} where \code{dataset_id} is
the id of an available dataset. Run \code{print_available_datasets()}
for obtaining the id of all available datasets.
The instantiated ProcessedQuant object needs to go through some
function calls, i.e, \code{fetch_quant()}, \code{decompress_quant()}
and \code{load_quant()} to become a complete ProcessedQuant class,
in which the \code{sce} slot contains the
\code{SingleCellExperiment} object of the quantification result
of the dataset. information of a specific dataset.}
\usage{
ProcessedQuant(dataset_id)
}
\arguments{
\item{dataset_id}{the id of an available dataset.}
}
\description{
The ProcessedQuant class
The ProcessedQuant class is designed to represent the details of
preprocessed dataset from the
\href{https://github.com/COMBINE-lab/10x-requant}{10x-requant}
project. To initialize a ProcessedQuant class, run
\code{ProcessedQuant(dataset_id)} where \code{dataset_id} is
the id of an available dataset. Run \code{print_available_datasets()}
for obtaining the id of all available datasets.
The instantiated ProcessedQuant object needs to go through some
function calls, i.e, \code{fetch_quant()}, \code{decompress_quant()}
and \code{load_quant()} to become a complete ProcessedQuant class,
in which the \code{sce} slot contains the
\code{SingleCellExperiment} object of the quantification result
of the dataset. information of a specific dataset.
}
