% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessedQuant.R
\name{load_quant}
\alias{load_quant}
\title{load the fetched quantification result of a
specific dataset as a SingleCellExperiment object
and store it in the sce slot of the returned
 processed_quant list as the sce slot.
This function must be run after \code{decompress_quant()}}
\usage{
load_quant(
  processed_quant,
  output_format = "scRNA",
  nonzero = FALSE,
  force = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{processed_quant}{a \code{ProcessedQuant} class object
recording the details of
a dataset.}

\item{output_format}{the format of the returned SingleCellExperiment
object. It will be passed to \code{\link[fishpond]{loadFry}}
as the \code{outputFormat} parameter.}

\item{nonzero}{It will be passed to \code{\link[fishpond]{loadFry}}
as the \code{nonzero} parameter.}

\item{force}{logic whether to proceed if the sce slot exists.}

\item{quiet}{logical whether to display no messages. Default is set as FALSE.}
}
\value{
A ProcessedQuant class object with a updated \code{sce} slot.
}
\description{
load the fetched quantification result of a
specific dataset as a SingleCellExperiment object
and store it in the sce slot of the returned
 processed_quant list as the sce slot.
This function must be run after \code{decompress_quant()}
}
